+++ assign01/matrix.h
@@ -12,7 +12,7 @@ class Matrix
 
 private:
     std::string name;
-    Eigen::MatrixXf mat;
+    Eigen::MatrixXd mat;
     int nrow;
     int ncol;
 
@@ -20,7 +20,7 @@ public:
     Matrix();
     Matrix(int dim);
     Matrix(int nrow, int ncol);
-    Matrix(Eigen::MatrixXf mat) : name("taylor"), mat(mat),
+    Matrix(Eigen::MatrixXd mat) : name("taylor"), mat(mat),
                                   nrow(static_cast<int>(mat.rows())), ncol(static_cast<int>(mat.cols())) {}
     Matrix(float p0, float p1, float p2, float p3, float p4, float p5, float p6, float p7, float p8);
     Matrix(float p0, float p1, float p2);
@@ -40,7 +40,7 @@ public:
     void assign(int row, int col, float val) { mat(row, col) = val; }
 
     std::string get_name() const { return name; }
-    Eigen::MatrixXf get() const { return mat; }
+    Eigen::MatrixXd get() const { return mat; }
     float get_pos(int row, int col) const { return mat(row, col); }
     int get_row() const { return nrow; }
     int get_col() const { return ncol; }
